+++ xpdacq/999-load2.py
@@ -1,8 +1,7 @@
 import os
-from pathlib import Path
-
 from bluesky.callbacks.zmq import Publisher
 from databroker.v2 import temp
+from pathlib import Path
 from xpdsim import xpd_pe1c, shctl1, cs700, ring_current, fb
 
 from xpdacq.beamtime import ScanPlan, Sample, ct, Tramp, Tlist, tseries